change the order of instructions in BaseView::refresh
The positionComponents() template method was taking place
before the wresize() and wmove() methods. This may result
in problems with incorrect information, in case when
the BaseView's subclass needs to access the main window's
current dimensions or position which are only available
once the above two methods have been called. That's why
positionComponents() method needs to follow the
wresize() and wmove() methods.