Remove dependency of builds on Linux container
Executing cmake (CMakeLists.txt) and building the executable were to a
certain extent dependent on being run in Docker container
preconfigured with Ubuntu distro of Linux. Project compilation failed
on OSs other than Linux, which posed a problem when one needed to debug
the built executable on their native platform.