Build a static lib out of generated proto files
The generated proto classes (grpc & protobuf) made the compiler to spit
out warnings which cannot be fixed due to the files being generated.
Suppressing the warnings with #pragma's does not work either as the
pragmas would have to be added also in some of the generated files.
Solution:
Build a static library out of the generated proto/grpc files and when
building this library, disable all warnings.
- split the main CMakeLists.txt file into two. The secondary now
contains all things related to grpc/protobuf.