Create TextEditView for editing text.
This commit comes with some base functionality of a TextEdit. It's not
perfect, has it's limitations and potential bugs, but it works.
Does not support horizontal or vertical scrolling. Performs a word wrap
by default. The amount of text that one can type in the TextEdit is
limited by the view's size.