Make View interface offer readInput() and setKeyFilter() functions.
readInput() and setKeyFilter() functions can now be called on any View.
readInput() - the View should read input from the user (at least one
character)
setKeyFilter() - takes in a lambda function, a callback for the caller
to respond to the keys read in the readInput() function. The lambda
should return true if the key should not be processed by the View.

View::setKeyFilter() is implemented in the BaseView class.
View::readInput() has a default implementation in BaseView, but should
be overriden by child classes of BaseView if needed.